Make WindowsDirectoryWatcherStrategy a friend of FileImpl
The WindowsDirectoryWatcherStrategy makes use of
Poco::FileImpl::convertPath which is protected within both of the
windows based FileImpl implementations (WIN32 and WINCE).  In the
WIN32 case though WindowsDirectoryWatcherStrategy is a friend class of
FileImpl.  This commit brings the WINCE implementation into line with
the WIN32 implementation, making WindowsDirectoryWatcherStrategy a
friend class.